refactor code, fix percentages calculations and props
